Split model.rs into multiple files
